Signal --> BaseSignal = bugs due to metadata.BaseSignal attributes
