+++ hyperspy/_signals/signal2D.py
@@ -25,7 +25,7 @@ import matplotlib.pyplot as plt
 from hyperspy.defaults_parser import preferences
 from hyperspy.external.progressbar import progressbar
 from hyperspy.misc.math_tools import symmetrize, antisymmetrize
-from hyperspy.signal import Signal
+from hyperspy.signal_base import BaseSignal
 
 
 def shift_image(im, shift, interpolation_order=1, fill_value=np.nan):
@@ -499,7 +499,7 @@ class Signal2DTools(object):
                   right)
 
 
-class Signal2D(Signal,
+class Signal2D(BaseSignal,
                Signal2DTools,):
 
     """
